pyproject.toml
[project]
name = "manger"
version = "0.1.0"
description = "Production-grade Manga Translation application using Magi OCR"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "streamlit>=1.28.0",
    "loguru>=0.7.0",
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "pymupdf>=1.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]
magi = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.35.0",
]

[project.scripts]
manger = "manger.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/manger"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]